// src/schools/dto/school-response.dto.ts
export class SchoolResponseDto {
  id: string;
  name: string;
  description: string;
  adminId: string;
  logo: string;
  banner: string;
  address: string;
  phone: string;
  email: string;
  totalClasses: number;
  totalStudents: number;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}